perf: copy buffer subslices to result instead of pushing result per byte and reversing
For most sane scenarios, where size of the buffer is greater than the length of the line, the result will only contain one and at most two elements, making the flattening trivial.
At the same time, instead of pushing one element at a time, it allows us to copy a subslice of the buffer, which is very performant on modern architectures.